organize css import type
