--- bikespace_frontend/src/components/shared/index.ts
@@ -1,1 +1,0 @@
-export {BaseButton} from './base-button';
