+++ bikespace_frontend/src/components/submission/submission-header/submission-header.module.scss
@@ -10,7 +10,7 @@
   height: 50px;
   padding: 0 1rem;
 
-  .header-logo img {
+  .headerLogo img {
     width: 147px;
   }
