+++ bikespace_frontend/tsconfig.json
@@ -24,6 +24,6 @@
     "./gatsby-config.ts",
     "./plugins/**/*",
     "./__tests__/**/*",
-    "./declaration.d.ts"
+    "./src/types/*"
   ]
 }
